use uint32 instead of uint16 for rle containers
stays consistent with the array type containers. array and rle are using uint32
in case we want to change the size of roaring containers as an optimization.
This has been done in the past
https://github.com/pilosa/pilosa/commit/a8feb985e15823bd461b841f7897f2060cdf284f

The downside is that more storage is used, so we may want to convert to uint16
until we decide to play with that optimization again, but for now I'll keep the
rle implementation in sync with the array implementation.