bump MaxIdleConns(PerHost) like v0.7
This helps "connection reset" issues when the cluster is under high query load
from many clients.

This also brings in the default http transport options which were (mistakenly?)
removed by the TLS work. I'm not sure what the ramifications would be of using a
blank http.Transport{} instead of http.DefaultTransport, but it seems best to
avoid a sweeping change.