keep gometalinter happy
gometalinter doesn't interact well with modules, neither do
some of the linters. But `go mod vendor` creates a working
vendor tree, such that you can use gometalinter, as long as
you don't mind keeping the pilosa module in a directory that
happens to be shaped like an old $GOPATH, in a subdirectory
named `github.com/pilosa/pilosa`.

And explicitly turn off GO111MODULES in that case, while
turning it on the rest of the time so `go mod` works.

This is a transient phase, eventual plan is to phase out
the old style entirely, but right no, gometalinter is nice
enough to jump through some hoops for.