add staticcheck
staticcheck is a flexible linter with a lot of potentially
useful diagnostics.

The current version doesn't find any problems in our tree,
but be warned that staticcheck-next will add a lot of
lint and style issues; with that in mind, I'm setting it
to not cause CI to fail by default for now, until we've
checked it out more.