address things identified by staticcheck-next
The 2018.1 version of staticcheck adds a number of new
diagnostics, getting it significantly closer to being a
complete replacement for gometalinter, only also one which
is more aggressive about a few things.

Issues addressed:
	* capitalization/punctuation of error messages
	* corresponding tests
	* receiver names of methods in a couple of cases
	* function/variable naming (capitalization and underscores)

Not addressed:
	* unused constants in pql.peg.go, one of which may be an
	  actual grammar issue?
	* unused functions in lru/lru.go; they're intentionally
	  unexported, and have a //nolint that megacheck recognizes
	  but staticcheck doesn't

Added staticcheck.conf files to suppress some additional
warnings, like the suggestion that at least one file in every
package should have a package comment. That might be worth
fixing, but not yet.

Note that the Makefile won't currently get the 2018.1 release
of staticcheck, I'm just getting ready for it and using that
as an excuse to get a feel for the kinds of things it reports
on.