do fancier logic for bit imports
This is a WIP for improving performance of importing large numbers
of values. For the friendly case where the values are coming in
sequentially, and every value is set, this is a very noticeable
speedup. For the less friendly case where values are coming in
essentially randomly, it's still somewhat better in limited testing.

ToDo: Figure out about what the cutoff is where the naive solution
is faster. I'd guess there will be one, although it will probably
vary somewhat with what formats the containers are currently in,
because converting to bitmap containers is a measurable cost.