WIP fixing bitmap.Any and putting a stop to empty containers
in the midst of this, I realized that merging this with Seebs' container size
reduction work was going to be a nightmare, so I'm holding off until that's
merged. Will have a separate PR that just fixes the immediate bug.

The underlying issue is that when bits are removed from a bitmap, the container
structs containing those bits are not automatically removed. Those empty
containers can stick around, which coupled with my incorrect loop was causing an
infinite hang. TODO when iter.Next() returns false, it should probably update
the key and value to 0, nil so that calls to iter.Value() don't keep returning a
real container.

I've added a number of checks around the code to try to stop empty containers
from being left in Bitmaps. A few parts of the Containers interface make this a
bit difficult though — like GetOrCreate and PutContainerValues (seriously, that
method is really strange).

Currently there is a panic in the tests after having added the nil checks in
Containers.Put.