WIP on simplifying fragment.go
use importPositions wherever possible. currently setBit and clearBit are
implemented this way. Realized that changing setValue will conflict with the
unbounded BSI work, so deferring this. Currently a test that checks stats is
broken.

setBit contains code which updates the maxRowID—importPositions currently
doesn't do that, so that may be a bug.